Extracts RenderEmoji as functional component
