+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import twemoji from 'twemoji'
 
+import DisplayEmoji from './assets/DisplayEmoji';
 import { Emojis } from './assets/emojis.js';
 import './App.css';
 
@@ -15,33 +15,15 @@ class App extends Component {
     });
   };
 
-  displayEmoji() {
-    const selectedEmoji = Emojis[this.state.index];
-    const twemojiMarkup = {
-      __html: twemoji.parse(selectedEmoji.emoji)
-    };
-
-    return (
-      <div className="test">
-        <div>emoji: {selectedEmoji.emoji}</div>
-        <div>codepoint: {selectedEmoji.codepoint}</div>
-        <div>parsed codepoint: {String.fromCodePoint(selectedEmoji.codepoint)}</div>
-        <div>twemoji: <div dangerouslySetInnerHTML={twemojiMarkup}/></div>
-        <div>index: {this.state.index}</div>
-      </div>
-    )
-  }
-
   render() {
     return (
       <div className="App">
         <h2>Welcome to Emoji Routlette</h2>
         <button onClick={this.getRandomEmoji}>Get new emoji</button>
-        {
-          (this.state.index < Emojis.length)
-            ? this.displayEmoji()
-            : <div>No emoji for index: {this.state.index}</div>
-        }
+        {(this.state.index < Emojis.length)
+          ? <DisplayEmoji index={this.state.index} />
+          : <div>No emoji for index: {this.state.index}</div>}
+
       </div>
     );
   }
